requirement.txt
amqp==2.5.2
backcall==0.1.0
billiard==3.6.3.0
boto3==1.12.39
botocore==1.15.39
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.2
Django==2.2
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.15.2
idna==2.9
importlib-metadata==1.6.0
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
jmespath==0.9.5
kombu==4.6.8
Markdown==3.2.1
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
pika==1.1.0
pkg-resources==0.0.0
prompt-toolkit==3.0.4
ptyprocess==0.6.0
py==1.8.1
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
retry==0.9.2
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.8
zipp==3.1.0
